requirements.txt
pandas
Pillow
matplotlib
datetime
yfinance
streamlit